fixed Tiled project directory shenanigans.
Turns out, if you make a project in Tiled, that project directory must be where you keep EVERYTHING for that tilemap. So that directory needs to be in your actual game project, can't just drag and drop files as needed because there are dependencies that it looks for within that Tiled project folder (which won't exist if you drag and drop stuff elsewhere). Lovely.